backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
pydantic[email]==2.8.2
slowapi==0.1.9
